Fix d2i_ECPKParameters_fp and i2d_ECPKParameters_fp macroes
These functions are part of the public API but we don't have tests
covering their usage.
They are actually implemented as macroes and the absence of tests has
caused them to fall out-of-sync with the latest changes to ASN1 related
functions and cause compilation warnings.

This commit fixes the public headers to reflect these changes.

Fixes #12443